MYPAGE >> 개인정보 수정 >> 기존프로필 있을때와 없을때의 미리보기...& 파일을 선택해야만 기존의 파일은 삭제하는 형식으로 적용

+++ src/main/java/com/sbs/hsb/ex1/service/MemberService.java
@@ -145,9 +145,12 @@ public class MemberService {
 	// 회원 개인 정보 수정 닉네임 이메일 프로필
 	public int doMemberModify(Map<String, Object> param) {
 		memberDao.doMemberModify(param);// 개인 정보 수정
-	
+
 		int id = Util.getAsInt(param.get("id"));
-		fileDao.deleteFiles("member", id);// 기존 회원 프로필 삭제
+		
+		if(param.get("fileIdsStr").toString().length() != 0) { // input한 이미지가 있을 경우에만
+			fileDao.deleteFiles("member", id);// 기존 회원 프로필 삭제
+		}
 
 		
 		String fileIdsStr = (String) param.get("fileIdsStr");
